Fixed a bug that had been introduced with the cycles: beacuse the number of seconds since the start of the epoch requires a certain precision in the floating point values, storing the time of death or time to depart as a float was not sufficient any more. This basically hints at a deeper problem with storing absolute times. Maybe everything should be converted to relative times.

+++ source/shot.h
@@ -38,7 +38,7 @@ public:
 	// setters
 	void SetDamage(float Damage);
 	void SetShooter(Object * Shooter);
-	void SetTimeOfDeath(float TimeOfDeath);
+	void SetTimeOfDeath(double TimeOfDeath);
 	void SetVelocity(const Vector3f & Velocity);
 private:
 	// modifers
@@ -73,7 +73,7 @@ inline void Shot::SetDamage(float Damage)
 	_Damage = Damage;
 }
 
-inline void Shot::SetTimeOfDeath(float TimeOfDeath)
+inline void Shot::SetTimeOfDeath(double TimeOfDeath)
 {
 	_TimeOfDeath = TimeOfDeath;
 }
